e2e framework: move kubectl and pod helper code
