+++ test/e2e/common/node/configmap.go
@@ -26,6 +26,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2etodopod "k8s.io/kubernetes/test/e2e/framework/todo/pod"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 	admissionapi "k8s.io/pod-security-admission/api"
 
@@ -79,7 +80,7 @@ var _ = SIGDescribe("ConfigMap", func() {
 			},
 		}
 
-		f.TestContainerOutput("consume configMaps", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "consume configMaps", pod, 0, []string{
 			"CONFIG_DATA_1=value-1",
 		})
 	})
@@ -123,7 +124,7 @@ var _ = SIGDescribe("ConfigMap", func() {
 			},
 		}
 
-		f.TestContainerOutput("consume configMaps", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "consume configMaps", pod, 0, []string{
 			"data-1=value-1", "data-2=value-2", "data-3=value-3",
 			"p-data-1=value-1", "p-data-2=value-2", "p-data-3=value-3",
 		})
