+++ test/e2e/common/storage/downwardapi_volume.go
@@ -27,6 +27,7 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
+	e2etodopod "k8s.io/kubernetes/test/e2e/framework/todo/pod"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 	admissionapi "k8s.io/pod-security-admission/api"
 
@@ -53,7 +54,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumePodForSimpleTest(podName, "/etc/podinfo/podname")
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("%s\n", podName),
 		})
 	})
@@ -69,7 +70,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		defaultMode := int32(0400)
 		pod := downwardAPIVolumePodForModeTest(podName, "/etc/podinfo/podname", nil, &defaultMode)
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			"mode of file \"/etc/podinfo/podname\": -r--------",
 		})
 	})
@@ -85,7 +86,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		mode := int32(0400)
 		pod := downwardAPIVolumePodForModeTest(podName, "/etc/podinfo/podname", &mode, nil)
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			"mode of file \"/etc/podinfo/podname\": -r--------",
 		})
 	})
@@ -100,7 +101,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 			FSGroup: &gid,
 		}
 		setPodNonRootUser(pod)
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("%s\n", podName),
 		})
 	})
@@ -116,7 +117,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 			FSGroup: &gid,
 		}
 		setPodNonRootUser(pod)
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			"mode of file \"/etc/podinfo/podname\": -r--r-----",
 		})
 	})
@@ -193,7 +194,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/podinfo/cpu_limit")
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("2\n"),
 		})
 	})
@@ -207,7 +208,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/podinfo/memory_limit")
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("67108864\n"),
 		})
 	})
@@ -221,7 +222,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/podinfo/cpu_request")
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("1\n"),
 		})
 	})
@@ -235,7 +236,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForContainerResources(podName, "/etc/podinfo/memory_request")
 
-		f.TestContainerOutput("downward API volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "downward API volume plugin", pod, 0, []string{
 			fmt.Sprintf("33554432\n"),
 		})
 	})
@@ -249,7 +250,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForDefaultContainerResources(podName, "/etc/podinfo/cpu_limit")
 
-		f.TestContainerOutputRegexp("downward API volume plugin", pod, 0, []string{"[1-9]"})
+		e2etodopod.TestContainerOutputRegexp(f, "downward API volume plugin", pod, 0, []string{"[1-9]"})
 	})
 
 	/*
@@ -261,7 +262,7 @@ var _ = SIGDescribe("Downward API volume", func() {
 		podName := "downwardapi-volume-" + string(uuid.NewUUID())
 		pod := downwardAPIVolumeForDefaultContainerResources(podName, "/etc/podinfo/memory_limit")
 
-		f.TestContainerOutputRegexp("downward API volume plugin", pod, 0, []string{"[1-9]"})
+		e2etodopod.TestContainerOutputRegexp(f, "downward API volume plugin", pod, 0, []string{"[1-9]"})
 	})
 })
