+++ test/e2e/common/storage/projected_combined.go
@@ -24,6 +24,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2etodopod "k8s.io/kubernetes/test/e2e/framework/todo/pod"
 	imageutils "k8s.io/kubernetes/test/utils/image"
 	admissionapi "k8s.io/pod-security-admission/api"
 
@@ -88,7 +89,7 @@ var _ = SIGDescribe("Projected combined", func() {
 				},
 			},
 		}
-		f.TestContainerOutput("Check all projections for projected volume plugin", pod, 0, []string{
+		e2etodopod.TestContainerOutput(f, "Check all projections for projected volume plugin", pod, 0, []string{
 			podName,
 			"secret-value-1",
 			"configmap-value-1",
