+++ test/e2e/framework/nodes_util.go
@@ -17,9 +17,6 @@ limitations under the License.
 package framework
 
 import (
-	"fmt"
-	"os"
-	"path"
 	"sync"
 	"time"
 
@@ -34,38 +31,6 @@ import (
 	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 )
 
-const etcdImage = "3.5.5-0"
-
-// EtcdUpgrade upgrades etcd on GCE.
-func EtcdUpgrade(targetStorage, targetVersion string) error {
-	switch TestContext.Provider {
-	case "gce":
-		return etcdUpgradeGCE(targetStorage, targetVersion)
-	default:
-		return fmt.Errorf("EtcdUpgrade() is not implemented for provider %s", TestContext.Provider)
-	}
-}
-
-func etcdUpgradeGCE(targetStorage, targetVersion string) error {
-	env := append(
-		os.Environ(),
-		"TEST_ETCD_VERSION="+targetVersion,
-		"STORAGE_BACKEND="+targetStorage,
-		"TEST_ETCD_IMAGE="+etcdImage)
-
-	_, _, err := RunCmdEnv(env, GCEUpgradeScript(), "-l", "-M")
-	return err
-}
-
-// LocationParamGKE returns parameter related to location for gcloud command.
-func LocationParamGKE() string {
-	if TestContext.CloudConfig.MultiMaster {
-		// GKE Regional Clusters are being tested.
-		return fmt.Sprintf("--region=%s", TestContext.CloudConfig.Region)
-	}
-	return fmt.Sprintf("--zone=%s", TestContext.CloudConfig.Zone)
-}
-
 // AppendContainerCommandGroupIfNeeded returns container command group parameter if necessary.
 func AppendContainerCommandGroupIfNeeded(args []string) []string {
 	if TestContext.CloudConfig.Region != "" {
@@ -73,55 +38,6 @@ func AppendContainerCommandGroupIfNeeded(args []string) []string {
 		return append([]string{"beta"}, args...)
 	}
 	return args
-}
-
-// MasterUpgradeGKE upgrades master node to the specified version on GKE.
-func MasterUpgradeGKE(namespace string, v string) error {
-	Logf("Upgrading master to %q", v)
-	args := []string{
-		"container",
-		"clusters",
-		fmt.Sprintf("--project=%s", TestContext.CloudConfig.ProjectID),
-		LocationParamGKE(),
-		"upgrade",
-		TestContext.CloudConfig.Cluster,
-		"--master",
-		fmt.Sprintf("--cluster-version=%s", v),
-		"--quiet",
-	}
-	_, _, err := RunCmd("gcloud", AppendContainerCommandGroupIfNeeded(args)...)
-	if err != nil {
-		return err
-	}
-
-	WaitForSSHTunnels(namespace)
-
-	return nil
-}
-
-// GCEUpgradeScript returns path of script for upgrading on GCE.
-func GCEUpgradeScript() string {
-	if len(TestContext.GCEUpgradeScript) == 0 {
-		return path.Join(TestContext.RepoRoot, "cluster/gce/upgrade.sh")
-	}
-	return TestContext.GCEUpgradeScript
-}
-
-// WaitForSSHTunnels waits for establishing SSH tunnel to busybox pod.
-func WaitForSSHTunnels(namespace string) {
-	Logf("Waiting for SSH tunnels to establish")
-	RunKubectl(namespace, "run", "ssh-tunnel-test",
-		"--image=busybox",
-		"--restart=Never",
-		"--command", "--",
-		"echo", "Hello")
-	defer RunKubectl(namespace, "delete", "pod", "ssh-tunnel-test")
-
-	// allow up to a minute for new ssh tunnels to establish
-	wait.PollImmediate(5*time.Second, time.Minute, func() (bool, error) {
-		_, err := RunKubectl(namespace, "logs", "ssh-tunnel-test")
-		return err == nil, nil
-	})
 }
 
 // NodeKiller is a utility to simulate node failures.
